change precision

+++ subtractAbs.m
@@ -1,3 +1,3 @@
 function out = subtractAbs(a,b)
-    out = abs(int8(a) - int8(b)); 
+    out = abs(int32(a) - int32(b)); 
 end
